[Add] Independent energy gain and loss in plot
The `overlap_ei_frac` yaml key previously required a single scalar and was then used to plot the sample-detector paths for neutrons with E=(1+f)*Ei and E=(1-f)*Ei. In some cases it may be useful to consider energy-gain and energy-loss neutrons with different fractional values so the key has been extended to accept a list of (at least) two numbers of the form '[g l]' where 'g' is the energy gain fraction and 'l' the energy loss fraction. To support typical notation 'g' is allowed to be negative. Both 'g' and 'l' can take any value that `float()` can convert, e.g., '[-inf 0.9]' is valid.