[Refactor] two 'phase' YAML keys
- The YAML keywords 'isPhaseIndependent' and 'defaultPhase' are used, collectively, to indicate whether a chopper has a standard ('isPhaseIndependent: False' or not present) phase relationship with the target incident energy; and in the case where something special is to be done whether the chopper should be phased relative to the source independent of selected incident energy ('defaultPhase: [non-string-value]') or if a specific disc slot should be phased to coincide with the focus energy ('defaultPhase: [string]').
We now want to support a third situation where a constant phase *offset* can be specified relative to the focus incident energy arrival time instead of the source repetition time (infinite energy arrival time).
- Previously the two keywords were reduced from per-chopper lists of {isPhaseIndependent: (True, False), defaultPhase: (None, [string], [not string])} to a list of indices which should be treated specially (isPhaseIndependent == True) and the non-None elements of 'defaultPhase' indicating *how* they should be handled.
Now the per-chopper lists are kept intact. If 'isPhaseIndependent' is True the chopper is phased relative to the source, otherwise it is phased relative to the incident energy. A 'defaultPhase' of None means no phase offset and the (arbitrarily) first slot should be used; a string 'defaultPhase' is used to select which slot should be phased (no phase offset), and any other value is interpreted as a floating point number phase offset.
- This change increases the number of settings from three to five:
  + phase at E_i
  + phase offset from the source by a value
  + phase at E_i through a selected slot
  + [new] phase offset from E_i by a value
  + [new] phase to the source through a selected slot
the last of which (probably) should not be used and, in any case, requires a change to the MAPS and MARI yaml files to set 'isPhaseIndependent' to False.